moved meter and button to scream component, signup component now classless
